internal/servers/worker: shut down connections on shutdown or reload
This adds functionality to ensure that connections are closed on a
server when the server is shut down or reloaded.

The behavior is as follows:

* Shutdown: Connections are closed, and then the shutdown process waits
for the next report to the controller to go through, waiting the
grace period that was added in #1330. The rest of shutdown is then
proceeded with. Listener shutdown has been moved to the start of the
process to ensure that new connections can't be made while we are
shutting down.

* Reload: Connections are simply closed. This has been added next to the
reload process that also processes tags.

The commit also includes some refactoring to tests so that we could use
some of the new E2E testing code to test session connections during
shutdown and reload.